# FIX : mmmove, strlcpy. + ADD : memchr

+++ ft_strlcpy.c
@@ -6,9 +6,11 @@
 /*   By: edetoh <edetoh@student.42lehavre.fr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/28 18:37:49 by edetoh            #+#    #+#             */
-/*   Updated: 2024/10/15 14:41:51 by edetoh           ###   ########.fr       */
+/*   Updated: 2024/10/18 10:44:44 by edetoh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
+#include "libft.h"
 
 /*
  * La fonction ft_strlcpy copie 'size' caractères de la chaîne
@@ -23,10 +25,10 @@
    en évitant les débordements de tampon.
  */
 
-unsigned int	ft_strlcpy(char *dest, char *src, unsigned int size)
+size_t	ft_strlcpy(char *dest, char *src, unsigned int size)
 {
-	unsigned int	count;
-	unsigned int	i;
+	size_t	count;
+	size_t	i;
 
 	count = 0;
 	while (src[count] != '\0')
@@ -34,7 +36,7 @@ unsigned int	ft_strlcpy(char *dest, char *src, unsigned int size)
 	if (size == 0)
 		return (count);
 	i = 0;
-	while (src[i] != '\0' && i < (size - 1))
+	while (src[i] != '\0' && i < ((size_t)size - 1))
 	{
 		dest[i] = src[i];
 		++i;
@@ -43,13 +45,17 @@ unsigned int	ft_strlcpy(char *dest, char *src, unsigned int size)
 	return (count);
 }
 
+// #include  <string.h>
+// #include  <stdio.h>
+// #include  <unistd.h>
+
 // int main(void)
 // {
 // 	char src[] = "Hello World!";
 // 	char dest[6];
 // 	int i;
 
-// 	i = ft_strlcpy(dest, src, 6);
+// 	i = ft_strlcpy(dest, src, 12);
 // 	printf("src: %s\n", src);
 // 	printf("dest: %s\n", dest);
 // 	printf("i: %d\n", i);
